requirements.txt
allure-pytest==2.8.17
jsonpath==0.82
loguru==0.5.1
pytest==6.0.1
PyYAML==5.3.1
requests==2.24.0
xlrd==1.2.0
yagmail==0.11.224
PyMySQL==0.10.1
pytest-rerunfailures==9.1.1
paramiko==2.7.2
pycryptodome==3.10.1